* More renaming in the editor sources. Fixed handling the SPACE key.
