+++ Neo/NeoEditor/MainWindow/MainWindow.cxx
@@ -1,11 +1,13 @@
 // generated by Fast Light User Interface Designer (fluid) version 1.0302
 
-#include "Translator.h"
-#include "MainWindow.h"
-#include "Callbacks.h"
+#include "../MainWindow/MainWindow.h"
+
 #include <string>
 #include <FL/Fl_Color_Chooser.H>
 #include <MeshSave.h>
+
+#include "../MainWindow/Callbacks.h"
+#include "../MainWindow/Translator.h"
 using namespace Neo;
 
 unsigned char EditorWindow::menu_menu_bar_i18n_done = 0;
