+++ Neo/NeoEditor/MainWindow/MainWindow.h
@@ -41,8 +41,8 @@ extern void add_group_callback(Fl_Menu_*, void*);
 extern void plugin_console_callback(Fl_Menu_*, void*);
 extern void check_for_updates_callback(Fl_Menu_*, void*);
 extern void about_menu_callback(Fl_Menu_*, void*);
-#include "GLBox.h"
-#include "DnDTree.h"
+#include "../MainWindow/GLBox.h"
+#include "../MainWindow/DnDTree.h"
 extern void scene_tree_callback(DnDTree*, long);
 #include <FL/Fl_Group.H>
 #include <FL/Fl_Round_Button.H>
