+++ Neo/NeoEditor/MainWindow/MainWindow.po
