+++ Neo/NeoEditor/MainWindow/Translator.cpp
@@ -17,25 +17,43 @@
 //
 //========================================================================
 
-#ifndef TRANSLATOR_H
-#define TRANSLATOR_H
+#include "../MainWindow/Translator.h"
 
-#include "ini.h"
+#include <NeoEngine.h>
 
-const char* tr(const char* key);
-
-class Translator
+const char* tr(const char* key)
 {
-    INI::Parser* m_parser;
-    char m_langFile[255];
+    return Translator::getInstance()->translate(key);
+}
 
-public:
-    static Translator* getInstance() { static Translator m_tr; return &m_tr; }
+void Translator::loadTranslation(const char* file)
+{
+    if(!isFileExist(file))
+		return;
 
-    Translator(): m_parser(NULL) {}
-    void loadTranslation(const char* file);
-    const char* translate(const char* key);
-    const char* getLanguageFile() { return m_langFile; }
-};
+    SAFE_DELETE(m_parser);
+    m_parser = new INI::Parser(file);
 
+#ifndef WIN32
+    const char* substr = strrchr(file, '/');
+#else
+    const char* substr = strrchr(file, '\\');
 #endif
+
+    if(substr == NULL)
+        substr = file;
+
+    strcpy(m_langFile, ++substr);
+}
+
+const char* Translator::translate(const char* key)
+{
+    if(!m_parser)
+        return key;
+
+    std::string value = m_parser->top()("Keys")[key];
+    if(value.empty())
+        return key;
+
+    return value.c_str();
+}
